pyproject.toml
[build-system]
requires = [
    "setuptools >= 57.0.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "tatc"
description = "Tradespace Analysis Toolkit for Constellations (TAT-C)"
authors = [
    {"name" = "Paul T. Grogan", "email" = "paul.grogan@asu.edu"}
]
readme = "README.md"
requires-python = ">=3.8, <3.11"
license = {text = "BSD License"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta"
]
dependencies = [
    "geopandas",
    "numba",
    "numpy",
    "pandas",
    "pydantic",
    "shapely",
    "skyfield",
    "sgp4"
]
dynamic = ["version"]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "black[jupyter] ~= 23.1.0",
    "coverage ~= 7.2.1",
    "pylint ~= 2.16.3",
    "pylint-pydantic ~= 0.1.7",
    "pytest ~= 7.1.3"
]
docs = [
    "autodoc_pydantic ~= 1.8.0",
    "docutils ~= 0.18.1",
    "ipython ~= 8.12.0",
    "myst-parser ~= 0.19.1",
    "nbsphinx ~= 0.9.1",
    "sphinx ~= 6.0",
    "sphinx-rtd-theme ~= 1.2.0"
]
examples = [
    "contextily",
    "ipykernel"
]
app = [
    "aiofiles ~= 23.1.0",
    "aiohttp ~= 3.8.4",
    "aiosqlite ~= 0.18.0",
    "celery[librabbitmq,redis] ~= 5.3.0",
    "fastapi ~= 0.95.0",
    "fastapi-users[sqlalchemy] ~= 10.4.1",
    "fastapi-users-db-sqlalchemy < 5.0.0",
    "fastapi-utils ~= 0.2.1",
    "geojson-pydantic ~= 0.5.0",
    "msgpack-asgi ~= 1.1.0",
    "python-dotenv ~= 1.0.0",
    "tzlocal ~= 4.3",
    "uvicorn[standard] ~= 0.21.1"
]

[tool.setuptools.dynamic]
version = {attr = "tatc.__version__"}